+++ spec/lib/graphql_rails/attributes/input_attribute_spec.rb
@@ -204,6 +204,16 @@ module GraphqlRails
           end
         end
 
+        context 'when input is deprecated' do
+          before do
+            attribute.deprecated('Use something else')
+          end
+
+          it 'builds field with a given deprecation reason' do
+            expect(input_argument_options).to include(deprecation_reason: 'Use something else')
+          end
+        end
+
         context 'when input has default value' do
           before do
             attribute.default_value('some value')